Final SK build (hopefully)
- Erased part of the snailking_build body that was showing above the
head in some states. This had the unfortunate disadvantage of breaking
the seamless alignment in the "mating call" state (we can't use it).
- Tons of changes on snailking_death02_build. There's one or two frames
just before the end of the anim where the body is out of place, but this
is such a fast-framed state that it probably won't be a problem.